Internalize run and result data collection for given month.
Make merging and collecting of run and result data for a given month
the collect_data method in the PbenchCombinedDataCollection class.
Make the es_data_gen function also a method in this class. Rename
merge_sos_and_perf_parallel to aggregate_data. aggregate_data now
only gives the collection object a month to collect data on and all
processing happens inside the Class object.
